Moved Renderer trait and its impls into separate source file
